+++ core/RefreshContextModel.js
@@ -4,7 +4,7 @@
  *
  * Copyright © 2018 Extremely Heavy Industries Inc.
  */
-import {HoistModel, XH} from '@xh/hoist/core';
+import {HoistModel, managed}  from '@xh/hoist/core';
 import {allSettled, start} from '@xh/hoist/promise';
 import {PendingTaskModel} from '@xh/hoist/utils/async';
 import {throwIf} from '@xh/hoist/utils/js';
@@ -13,25 +13,27 @@ import {pull} from 'lodash';
 /**
  * Model to manage refreshing sections of the application, where "refreshing" refers to app-specific
  * actions to load and display updated data. Calling `refreshAsync()` on this model will cause it to
- * trigger a refresh on all linked models within its containing context.
+ * trigger a refresh on all linked models within the graphical hierarchy it contains.
  *
- * A global instance of this class is provided by the framework as `XH.refreshModel`. Apps can
- * create additional RefreshModels and pass them to nested RefreshView components to establish
- * distinct areas of the component hierarchy that can then be refreshed independently.
+ * A global instance of this class is provided by the framework as `XH.refreshContextModel`. Apps can
+ * create additional sub-contexts using a RefreshContextView.
  *
- * Note that TabContainer automatically establishes separate RefreshModels for its tabs and uses
- * these to implement efficient refresh handling of inactive / not-yet-rendered tabs. See the
+ * Note that TabContainer automatically establishes separate refresh contexts for its tabs and uses
+ * these to implement efficient refresh handling of inactive and not-yet-rendered tabs. See the
  * `refreshMode` configs on `TabContainerModel` and `TabModel` for more information on this common
  * use case.
  *
- * @see RefreshView
+ * @see RefreshContext
+ * @see RefreshContextView
  */
 @HoistModel
-export class RefreshModel {
+export class RefreshContextModel {
 
     targets = [];
 
+    @managed
     loadModel = new PendingTaskModel();
+
     lastRefreshRequested = null;
     lastRefreshCompleted = null;
 
@@ -55,7 +57,7 @@ export class RefreshModel {
      * @param {HoistModel} target
      */
     register(target) {
-        throwIf(!target.isHoistModel, 'Cannot register non-HoistModel with RefreshModel.');
+        throwIf(!target.isHoistModel, 'Cannot register non-HoistModel with RefreshContextModel.');
         const {targets} = this;
         if (!targets.includes(target)) targets.push(target);
     }
@@ -70,10 +72,5 @@ export class RefreshModel {
      */
     unregister(target) {
         pull(this.targets, target);
-    }
-
-    destroy() {
-        this.targets = null;
-        XH.safeDestroy(this.loadModel);
     }
 }