+++ desktop/appcontainer/AppContainer.js
@@ -7,7 +7,7 @@
 
 import {Component} from 'react';
 import {observable, runInAction} from '@xh/hoist/mobx';
-import {HoistComponent, elemFactory, elem, AppState, XH} from '@xh/hoist/core';
+import {HoistComponent, elemFactory, elem, AppState, refreshContextView, XH} from '@xh/hoist/core';
 import {mask} from '@xh/hoist/desktop/cmp/mask';
 import {div, frame, vframe, viewport} from '@xh/hoist/cmp/layout';
 
@@ -28,7 +28,6 @@ import {AppContainerModel} from '@xh/hoist/core/appcontainer/AppContainerModel';
 import {StoreContextMenu} from '@xh/hoist/desktop/cmp/contextmenu';
 import {colChooser, ColChooserModel} from '@xh/hoist/desktop/cmp/grid';
 import {installDesktopImpls} from '@xh/hoist/dynamics/desktop';
-import {refreshView} from '@xh/hoist/cmp/refresh';
 
 installDesktopImpls({
     colChooser,
@@ -92,8 +91,8 @@ export class AppContainer extends Component {
                     vframe(
                         impersonationBar({model: model.impersonationBarModel}),
                         updateBar({model}),
-                        refreshView({
-                            model: model.refreshModel,
+                        refreshContextView({
+                            model: model.refreshContextModel,
                             item: frame(elem(XH.appSpec.componentClass, {model: XH.appModel}))
                         }),
                         versionBar({model})