+++ desktop/cmp/button/RefreshButton.js
@@ -17,7 +17,7 @@ import {withDefault} from '@xh/hoist/utils/js';
  * Convenience Button preconfigured for use as a trigger for a refresh operation.
  *
  * If a model is provided it will be directly refreshed.  Alternatively an onClick handler
- * may be provided.  If neither of these props are provided, the contextual RefreshModel for this
+ * may be provided.  If neither of these props are provided, the contextual RefreshContextModel for this
  * button will be used.
  */
 @HoistComponent