+++ desktop/cmp/tab/impl/TabRefreshContextModel.js
@@ -4,12 +4,12 @@
  *
  * Copyright © 2018 Extremely Heavy Industries Inc.
  */
-import {RefreshModel} from '@xh/hoist/core';
+import {RefreshContextModel} from '@xh/hoist/core';
 
 /**
  * @private
  */
-export class TabRefreshModel extends RefreshModel {
+export class TabRefreshContextModel extends RefreshContextModel {
 
     tabModel;