+++ mobile/appcontainer/AppContainer.js
@@ -7,11 +7,10 @@
 
 import {Component} from 'react';
 import {observable, runInAction} from '@xh/hoist/mobx';
-import {HoistComponent, elem, elemFactory, AppState, XH} from '@xh/hoist/core';
+import {HoistComponent, elem, elemFactory, refreshContextView, AppState, XH} from '@xh/hoist/core';
 import {div, frame, vframe, viewport} from '@xh/hoist/cmp/layout';
 import {mask} from '@xh/hoist/mobile/cmp/mask';
 import {menu} from '@xh/hoist/mobile/cmp/menu';
-import {refreshView} from '@xh/hoist/cmp/refresh';
 
 import {aboutDialog} from './AboutDialog';
 import {feedbackDialog} from './FeedbackDialog';
@@ -81,8 +80,8 @@ export class AppContainer extends Component {
                     vframe(
                         impersonationBar({model: model.impersonationBarModel}),
                         updateBar({model}),
-                        refreshView({
-                            model: model.refreshModel,
+                        refreshContextView({
+                            model: model.refreshContextModel,
                             item: frame(elem(XH.appSpec.componentClass, {model: XH.appModel}))
                         }),
                         versionBar({model}),